+++ src/Identity/Logitar.Identity.Domain/Users/Events/UserDeletedEvent.cs
@@ -4,8 +4,9 @@ namespace Logitar.Identity.Domain.Users.Events;
 
 public record UserDeletedEvent : DomainEvent
 {
-  public UserDeletedEvent()
+  public UserDeletedEvent(ActorId actorId = default)
   {
+    ActorId = actorId;
     IsDeleted = true;
   }
 }