+++ src/Identity/Logitar.Identity.Domain/Users/Events/UserEnabledEvent.cs
@@ -2,4 +2,10 @@
 
 namespace Logitar.Identity.Domain.Users.Events;
 
-public record UserEnabledEvent : DomainEvent;
+public record UserEnabledEvent : DomainEvent
+{
+  public UserEnabledEvent(ActorId actorId = default)
+  {
+    ActorId = actorId;
+  }
+}