+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created a RoleAggregate.
 - Created an UserAggregate.
 
+### Changed
+
+- Protected setters on aggregate metadata.
+
 ## [1.2.0] - 2023-08-06
 
 ### Added
@@ -28,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reorganized the solution directory structure.
 - Upgraded NuGet packages and fixed EventSourcing project dependencies.
 - Replaced DeleteAction by a nullable boolean.
-- Refactored AggregateRoot & DomainEvent.
+- Refactored AggregateRoot and DomainEvent.
 - Injecting IEventSerializer as a dependency.
 
 ## [1.1.0] - 2023-07-11